The verification form that is shown to the user and then verified.
* The administrator registers
* One of the present administrators, goes into the database and changes the verification
string for the new administrator and sends an email to the new administrator
containing the verification string to be entered.

On the edit profile page a link to the verify_administrator is shown, if the administrator
has not yet been verified.

The manual step of mailing can't be avoided, because, since only some people are allowed
to be verified admins, it's important to put a manual check.